+++ packages/hms_room_kit/lib/src/meeting/meeting_header.dart
@@ -42,7 +42,10 @@ class _MeetingHeaderState extends State<MeetingHeader> {
                 ? Container()
                 : HMSRoomLayout.roleLayoutData!.logo!.url!.contains("svg")
                     ? SvgPicture.network(
-                        HMSRoomLayout.roleLayoutData!.logo!.url!)
+                        HMSRoomLayout.roleLayoutData!.logo!.url!,
+                        height: 30,
+                        width: 30,
+                      )
                     : Image.network(
                         HMSRoomLayout.roleLayoutData!.logo!.url!,
                         height: 30,