+++ packages/hms_room_kit/lib/src/meeting/meeting_store.dart
@@ -604,6 +604,13 @@ class MeetingStore extends ChangeNotifier
       previousRole = localPeer?.role.name;
       if (isRaisedHand) {
         changeMetadata();
+      } else {
+        ///Setting the previous role
+        String value = isRaisedHand ? "true" : "false";
+        _hmsSDKInteractor.changeMetadata(
+            metadata:
+                "{\"isHandRaised\":$value,\"isBRBOn\":false,\"prevRole\":\"$previousRole\"}",
+            hmsActionResultListener: this);
       }
       HMSRoomLayout.resetLayout(hmsRoleChangeRequest.suggestedRole.name);
       currentRoleChangeRequest = null;
@@ -1184,6 +1191,11 @@ class MeetingStore extends ChangeNotifier
     peerTracks.clear();
     isRoomEnded = true;
     resetForegroundTaskAndOrientation();
+
+    ///Here we call the method passed by the user in HMSPrebuilt as a callback
+    if (Constant.onLeave != null) {
+      Constant.onLeave!();
+    }
     notifyListeners();
   }