+++ packages/hms_room_kit/lib/src/screen_controller.dart
@@ -24,7 +24,13 @@ class ScreenController extends StatefulWidget {
   ///For more details checkout the [HMSPrebuiltOptions] class
   final HMSPrebuiltOptions? options;
 
-  const ScreenController({super.key, required this.roomCode, this.options});
+  ///The callback for the leave room button
+  ///This function can be passed if you wish to perform some specific actions
+  ///in addition to leaving the room when the leave room button is pressed
+  final Function? onLeave;
+
+  const ScreenController(
+      {super.key, required this.roomCode, this.options, this.onLeave});
   @override
   State<ScreenController> createState() => _ScreenControllerState();
 }
@@ -42,6 +48,7 @@ class _ScreenControllerState extends State<ScreenController> {
     ///Setting the prebuilt options and roomCode
     Constant.prebuiltOptions = widget.options;
     Constant.roomCode = widget.roomCode;
+    Constant.onLeave = widget.onLeave;
 
     ///Here we set the endPoints if it's non-null
     if (widget.options?.endPoints != null) {