+++ packages/hms_room_kit/lib/src/widgets/bottom_sheets/more_settings_bottom_sheet.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:badges/badges.dart' as badge;
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:google_fonts/google_fonts.dart';
+import 'package:hms_room_kit/src/widgets/common_widgets/hms_text_style.dart';
 import 'package:hms_room_kit/src/layout_api/hms_theme_colors.dart';
 import 'package:hmssdk_flutter/hmssdk_flutter.dart';
 import 'package:hms_room_kit/src/common/app_color.dart';
@@ -11,7 +11,6 @@ import 'package:hms_room_kit/src/common/constants.dart';
 import 'package:hms_room_kit/src/common/utility_components.dart';
 import 'package:hms_room_kit/src/widgets/app_dialogs/stats_for_nerds.dart';
 import 'package:hms_room_kit/src/widgets/bottom_sheets/audio_settings_bottom_sheet.dart';
-import 'package:hms_room_kit/src/widgets/bottom_sheets/meeting_mode_bottom_sheet.dart';
 import 'package:hms_room_kit/src/widgets/bottom_sheets/participants_bottom_sheet.dart';
 import 'package:hms_room_kit/src/widgets/bottom_sheets/start_hls_bottom_sheet.dart';
 import 'package:hms_room_kit/src/widgets/common_widgets/share_link_option.dart';
@@ -144,7 +143,7 @@ class _MoreSettingsBottomSheetState extends State<MoreSettingsBottomSheet> {
                               Text(
                                 "Participants",
                                 semanticsLabel: "participants_button",
-                                style: GoogleFonts.inter(
+                                style: HMSTextStyle.setTextStyle(
                                     fontSize: 14,
                                     color: themeDefaultColor,
                                     letterSpacing: 0.25,
@@ -190,7 +189,7 @@ class _MoreSettingsBottomSheetState extends State<MoreSettingsBottomSheet> {
                               Text(
                                 "Audio Settings",
                                 semanticsLabel: "fl_audio_settings",
-                                style: GoogleFonts.inter(
+                                style: HMSTextStyle.setTextStyle(
                                     fontSize: 14,
                                     color: themeDefaultColor,
                                     letterSpacing: 0.25,
@@ -205,39 +204,6 @@ class _MoreSettingsBottomSheetState extends State<MoreSettingsBottomSheet> {
                   const SizedBox(
                     height: 10,
                   ),
-                  if (Constant.debugMode)
-                    ListTile(
-                      horizontalTitleGap: 2,
-                      onTap: () async {
-                        Navigator.pop(context);
-                        showModalBottomSheet(
-                          isScrollControlled: true,
-                          backgroundColor: HMSThemeColors.surfaceDim,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(20),
-                          ),
-                          context: context,
-                          builder: (ctx) => ChangeNotifierProvider.value(
-                              value: meetingStore,
-                              child: const MeetingModeBottomSheet()),
-                        );
-                      },
-                      contentPadding: EdgeInsets.zero,
-                      leading: SvgPicture.asset(
-                        "packages/hms_room_kit/lib/src/assets/icons/participants.svg",
-                        height: 20,
-                        width: 20,
-                        colorFilter: ColorFilter.mode(
-                            HMSThemeColors.onSurfaceHighEmphasis,
-                            BlendMode.srcIn),
-                      ),
-                      title: HMSSubheadingText(
-                        text: "Meeting mode",
-                        textColor: HMSThemeColors.onSurfaceHighEmphasis,
-                        letterSpacing: 0.15,
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
                   if (meetingStore.localPeer?.role.publishSettings?.allowed
                           .contains("screen") ??
                       false)
@@ -698,7 +664,7 @@ class _MoreSettingsBottomSheetState extends State<MoreSettingsBottomSheet> {
                   //     title: Text(
                   //       "Modify Notifications",
                   //       semanticsLabel: "fl_notification_setting",
-                  //       style: GoogleFonts.inter(
+                  //       style: HMSTextStyle.setTextStyle(
                   //           fontSize: 14,
                   //           colorFilter:  ColorFilter.mode(themeDefaultColor, BlendMode.srcIn),
                   //           letterSpacing: 0.25,