+++ packages/hms_room_kit/CHANGELOG.md
@@ -1,9 +1,21 @@
 # Latest Versions
 
-| Package           | Version                                                                                                   |
-| ------------------| --------------------------------------------------------------------------------------------------------- |
-| hms_room_kit      | [![Pub Version](https://img.shields.io/pub/v/hms_room_kit)](https://pub.dev/packages/hms_room_kit)        |
-| hmssdk_flutter    | [![Pub Version](https://img.shields.io/pub/v/hmssdk_flutter)](https://pub.dev/packages/hmssdk_flutter)    |
+| Package        | Version                                                                                                |
+| -------------- | ------------------------------------------------------------------------------------------------------ |
+| hms_room_kit   | [![Pub Version](https://img.shields.io/pub/v/hms_room_kit)](https://pub.dev/packages/hms_room_kit)     |
+| hmssdk_flutter | [![Pub Version](https://img.shields.io/pub/v/hmssdk_flutter)](https://pub.dev/packages/hmssdk_flutter) |
+
+## 1.0.2 - 2023-09-22
+
+### 🚀 Added
+
+- Added `onLeave` callback parameter in `HMSPrebuilt`
+
+### Changed
+
+- Bump minimum Flutter version to 3.10.0
+
+- Bug fixes and performance improvements
 
 ## 1.0.1 - 2023-09-15
 
@@ -22,7 +34,7 @@
 - Preview For Role
 - Edge To Edge HLS Player
 - Updated Bottom Sheets
-And much more. Take it for a spin!
+  And much more. Take it for a spin!
 
 ## 0.1.0 - 2023-08-18
 
@@ -32,7 +44,7 @@ And much more. Take it for a spin!
 - Active Speaker Sorting
 - Newly Designed Header and Footer
 - Inset Tile for Local Peer
-And much more. Take it for a spin!
+  And much more. Take it for a spin!
 
 ## 0.0.4 - 2023-08-01